Instagram API interface
I added a defines files that just has a dictionary full of useful information. A lot of the information needs to be obtained from the developer account, but you can access that at anytime. I added a couple of functions for the Instagram API but still need to figure out which permissions I need for my access token to get them.